fix(api): update remove pin int conversion and use fromMap instead of fromJson

+++ frontend-mobile-flutter/lib/features/auth/view_models/auth_view_model.dart
@@ -148,7 +148,7 @@ class AuthViewModel extends _$AuthViewModel {
 
     Either<AppFailure, UserModel> result = await _authRemoteRepository.login(
       email: emailOrPin.contains("@") ? emailOrPin : null,
-      pin: !emailOrPin.contains("@") ? emailOrPin as int : null,
+      pin: !emailOrPin.contains("@") ? emailOrPin : null,
       password: password,
     );
